test: fix integer size cast warnings
On 64-bit, casting a pointer to int results in a warning
("cast from pointer to integer of different size"). Fix
that by using uintptr_t as the integer type here.

If the offset were too big for unsigned int (which really
hopefully doesn't happen - who has structures that big?)
we should get an integer truncation warning on the init.

+++ include/usfstl/test.h
@@ -177,7 +177,7 @@ struct usfstl_testcase {
 	.testcases = cases,						\
 	.testcase_size = sizeof(cases[0]),				\
 	.testcase_count = sizeof(cases)/sizeof(cases[0]),		\
-	.testcase_generic_offset = (int)((char *)&((typeof(&cases[0]))0)->generic),
+	.testcase_generic_offset = (uintptr_t)((char *)&((typeof(&cases[0]))0)->generic),
 
 #define _USFSTL_ASSIGN_TESTCASES2(...)
 
@@ -416,7 +416,7 @@ static void *tn##_generator(const struct usfstl_test *t, unsigned int n);	\
 type(tn, extra, NO_CASES, .case_generator = tn##_generator,			\
 	.case_generator_has_generic = true,					\
 	.testcase_generic_offset =						\
-		(int)((char *)&((struct usfstl_code_testcase *)0)->generic), \
+		(uintptr_t)((char *)&((struct usfstl_code_testcase *)0)->generic),\
 	__VA_ARGS__)
 #define _USFSTL_CODE_TEST_INIT(type, tn, extra, ...)				\
 	__USFSTL_CODE_TEST_INIT(type, tn, extra, __VA_ARGS__)
